Fix for failure to stop and added VISCA speed units to on-screen display.
Added a workaround for a bug in some cameras (multiple brands, likely)
where the pan and tilt could randomly get stuck moving when they shouldn't
be and ignored stop commands.

At a high level, if the camera should be stopped, but is still moving then
after at most 50 milliseconds of moving, the software fudges the next
speed value and sending -1 instead of 0.

Here's where this gets weird.  The workaround itself isn't doing anything.
Adding the workaround code to detect whether the camera is still moving or
not appears to have actually prevented the cameras from exhibiting the bug,
so there's no way to know if the workaround itself actually works, because
it never gets triggered anymore.

This bug was *easily* reproduced before adding the workaround code.

This at least suggests that hammering the camera repeatedly with position
queries (the rate of queries was also increased from 10 per second to 100 per
second) is somehow making the firmware realize that the camera is moving
when it shouldn't be, and is stopping it.  I have no idea how that is
possible unless it is some race condition in the firmware that being under
high load is preventing.